feat: no separate matmul nor online softmax
