+++ CMakeLists.txt
@@ -183,7 +183,8 @@ ascendc_library(
   csrc/kernel/kernel_kda_gate_cumsum.cpp
   csrc/kernel/kernel_kda_wy.cpp
   csrc/kernel/kernel_gdn_wy_fast.cpp
-  csrc/kernel/kernel_kda_kkt.cpp)
+    csrc/kernel/kernel_kda_kkt.cpp
+    csrc/kernel/kernel_fa.cpp)
 
 if(USE_ASC_LANGUAGE)
   # pto_kernels_ops is linked with RUNPATH $ORIGIN/build/lib, so the kernel
